requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --resolver=backtracking requirements.in
#
appdirs==1.4.4
    # via fs
attrs==22.2.0
    # via
    #   -r requirements.in
    #   cattrs
    #   statmake
    #   ufolib2
booleanoperations==0.9.0
    # via ufo2ft
cattrs==22.2.0
    # via statmake
cffsubr==0.2.9.post1
    # via ufo2ft
cu2qu==1.6.7.post2
    # via ufo2ft
fontmath==0.9.3
    # via -r requirements.in
fonttools[lxml,ufo,unicode]==4.38.0
    # via
    #   -r requirements.in
    #   booleanoperations
    #   cffsubr
    #   cu2qu
    #   fontmath
    #   psautohint
    #   statmake
    #   ufo2ft
    #   ufolib2
fs==2.4.16
    # via fonttools
lxml==4.9.2
    # via fonttools
psautohint==2.4.0
    # via -r requirements.in
pyclipper==1.3.0.post4
    # via booleanoperations
six==1.16.0
    # via fs
skia-pathops==0.7.4
    # via -r requirements.in
statmake==0.6.0
    # via -r requirements.in
ufo2ft[cffsubr]==2.30.0
    # via -r requirements.in
ufolib2==0.14.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
